// frontend/interface/infinityhealth.interface.ts
export type MetricType = 'Weight' | 'Height' | 'BMI' | 'Water' | 'Sleep' | 'Steps';